CHEF-2601 : Package renamed
