+++ components/automate-cli/pkg/verifyserver/services/batchcheckservice/trigger/certificatechecktrigger/certificatechecktrigger.go
@@ -1,4 +1,4 @@
-package trigger
+package certificatechecktrigger
 
 import (
 	"encoding/json"
