+++ components/automate-cli/pkg/verifyserver/services/batchcheckservice/trigger/sshuseraccesschecktrigger/sshuseraccesschecktrigger_test.go
@@ -1,4 +1,4 @@
-package trigger
+package sshuseraccesschecktrigger
 
 import (
 	"encoding/json"
